+++ Resources/createSelfExtracting.zsh
@@ -6,8 +6,8 @@
 #
 # Updated by: Dan K. Snelson
 # For DDM OS Reminder v2.0.0+
-# Version: 2.1.0
-# Date: 13-Dec-2025
+# Version: 2.2.0b1
+# Date: 15-Dec-2025
 #
 # Creates a self-extracting, base64-encoded shell script from
 # the newest "ddm-os-reminder-assembled-*.zsh" file found in