+++ launchDaemonManagement.zsh
@@ -21,16 +21,9 @@
 #
 # HISTORY
 #
-# Version 2.1.0, 13-Dec-2025, Dan K. Snelson (@dan-snelson)
-#   - Added ability to use `titleMessageUpdateOrUpgrade:l` (Pull Request #26; thanks, @maxsundellacne!)
-#   - Added logic to hide `button2` based on `DaysBeforeDeadlineHidingButton2` (Pull Request #27; thanks, @maxsundellacne!)
-#   - Refactored `resetConfiguration` function to avoid errors when attempting to `chmod` non-existent files
-#   - Added warning for excessive uptime (configurable via `DaysOfExcessiveUptimeWarning` variable; Feature Request #28)
-#   - Added logic for when the reminder dialog is re-displayed after clicking the `infobutton` (based on if we're already hiding the secondary button; #31)
-#   - Streamline Deployment & Documentation (Feature Request #35)
-#   - Addressed Bugs #34 (thanks, @TechTrekkie!) and #36 (I. Blame. AI.)
-#   - Refactored `assemble.zsh` (thanks for the feedback, @Andrew!)
-#   - Added warning for low disk space (configurable via `minimumDiskFreePercentage` variable; Feature Request #39. (Thanks for the suggestion, @prgsenright!)
+# Version 2.2.0b1, 15-Dec-2025, Dan K. Snelson (@dan-snelson)
+#   - Addressed Feature Request: Intelligently display reminder dialog after rebooting #42
+#   - Added instructions for monitoring the client-side log
 #
 ####################################################################################################
 
@@ -45,7 +38,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.1.0"
+scriptVersion="2.2.0b1"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"
@@ -591,5 +584,9 @@ launchDaemonStatus
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 # Exit
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
+
+quitOut "Completed ${reverseDomainNameNotation}.${organizationScriptName} LaunchDaemon"
+quitOut "Monitor the client-side log via:"
+quitOut "tail -f ${scriptLog}"
 
 exit 0